+++ core/src/bytes/mod.rs
@@ -69,39 +69,39 @@ impl<F: Field> ByteChip<F> {
                     ByteOpcode::AND => {
                         let and = b & c;
                         col.and = F::from_canonical_u8(and);
-                        ByteLookupEvent::new(*opcode, and, 0, b, c)
+                        ByteLookupEvent::new(*opcode, and as u32, 0, b as u32, c as u32)
                     }
                     ByteOpcode::OR => {
                         let or = b | c;
                         col.or = F::from_canonical_u8(or);
-                        ByteLookupEvent::new(*opcode, or, 0, b, c)
+                        ByteLookupEvent::new(*opcode, or as u32, 0, b as u32, c as u32)
                     }
                     ByteOpcode::XOR => {
                         let xor = b ^ c;
                         col.xor = F::from_canonical_u8(xor);
-                        ByteLookupEvent::new(*opcode, xor, 0, b, c)
+                        ByteLookupEvent::new(*opcode, xor as u32, 0, b as u32, c as u32)
                     }
                     ByteOpcode::SLL => {
                         let sll = b << (c & 7);
                         col.sll = F::from_canonical_u8(sll);
-                        ByteLookupEvent::new(*opcode, sll, 0, b, c)
+                        ByteLookupEvent::new(*opcode, sll as u32, 0, b as u32, c as u32)
                     }
-                    ByteOpcode::Range => ByteLookupEvent::new(*opcode, 0, 0, b, c),
+                    ByteOpcode::Range => ByteLookupEvent::new(*opcode, 0, 0, b as u32, c as u32),
                     ByteOpcode::ShrCarry => {
                         let (res, carry) = shr_carry(b, c);
                         col.shr = F::from_canonical_u8(res);
                         col.shr_carry = F::from_canonical_u8(carry);
-                        ByteLookupEvent::new(*opcode, res, carry, b, c)
+                        ByteLookupEvent::new(*opcode, res as u32, carry as u32, b as u32, c as u32)
                     }
                     ByteOpcode::LTU => {
                         let ltu = b < c;
                         col.ltu = F::from_bool(ltu);
-                        ByteLookupEvent::new(*opcode, ltu as u8, 0, b, c)
+                        ByteLookupEvent::new(*opcode, ltu as u32, 0, b as u32, c as u32)
                     }
                     ByteOpcode::MSB => {
                         let msb = (b & 0b1000_0000) != 0;
                         col.msb = F::from_bool(msb);
-                        ByteLookupEvent::new(*opcode, msb as u8, 0, b, 0)
+                        ByteLookupEvent::new(*opcode, msb as u32, 0, b as u32, 0 as u32)
                     }
                 };
                 event_map.insert(event, (row_index, i));
